+++ src/UI/Inventory/InstanceRow.tsx
@@ -13,7 +13,12 @@ import {
   DateWithTooltip,
   IdWithCopy,
 } from "./Components";
-import { InstanceDetails, AttributesView } from "@/UI/ServiceInstanceDetails";
+import {
+  InstanceDetails,
+  AttributesView,
+  StatusView,
+} from "@/UI/ServiceInstanceDetails";
+import { InfoCircleIcon, ListIcon } from "@patternfly/react-icons";
 
 interface Props {
   row: Row;
@@ -54,13 +59,28 @@ export const InstanceRow: React.FC<Props> = ({
       <Td dataLabel={words("inventory.column.updatedAt")}>
         <DateWithTooltip date={row.updatedAt} />
       </Td>
-      <Td dataLabel={words("inventory.column.actions")}>{actions}</Td>
     </Tr>
     <Tr isExpanded={isExpanded} data-testid={`details_${row.id.short}`}>
       <Td colSpan={numberOfColumns}>
         <ExpandableRowContent>
           <InstanceDetails>
-            <AttributesView attributes={row.attributes} />
+            <StatusView
+              title={words("inventory.tabs.status")}
+              icon={<InfoCircleIcon />}
+              statusInfo={{
+                instanceId: row.id.full,
+                state: row.state,
+                version: row.version,
+                createdAt: row.createdAt.full,
+                updatedAt: row.updatedAt.full,
+                actions: actions,
+              }}
+            />
+            <AttributesView
+              attributes={row.attributes}
+              title={words("inventory.tabs.attributes")}
+              icon={<ListIcon />}
+            />
           </InstanceDetails>
         </ExpandableRowContent>
       </Td>